parted code of counting instance for contexts and entity definitions
