+++ ui/features/eu.esdihumboldt.hale.ui.feature/feature.xml
@@ -499,4 +499,11 @@
          version="0.0.0"
          unpack="false"/>
 
+   <plugin
+         id="eu.esdihumboldt.hale.ui.service.helper"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"
+         unpack="false"/>
+
 </feature>
