+++ ui/plugins/eu.esdihumboldt.hale.ui.service.helper/.settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,4 @@
+#Created from default preferences 14.11.2016 15:03:15
+#Mon Nov 14 15:03:15 CET 2016
+eclipse.preferences.version=1
+encoding/<project>=UTF-8
