+++ ui/plugins/eu.esdihumboldt.hale.ui.service.helper/build.properties
@@ -1,0 +1,4 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .
