+++ ui/plugins/eu.esdihumboldt.hale.ui/META-INF/MANIFEST.MF
@@ -100,6 +100,7 @@ Import-Package: com.google.common.base;version="1.6.0",
  eu.esdihumboldt.hale.ui.filter.extension,
  eu.esdihumboldt.hale.ui.geometry.service,
  eu.esdihumboldt.hale.ui.geometry.service.impl,
+ eu.esdihumboldt.hale.ui.service.helper.population,
  eu.esdihumboldt.hale.ui.util,
  eu.esdihumboldt.hale.ui.util.dialog,
  eu.esdihumboldt.hale.ui.util.handler,
