+++ ui/plugins/eu.esdihumboldt.hale.ui/src/eu/esdihumboldt/hale/ui/service/population/internal/PopulationServiceImpl.java
@@ -16,7 +16,6 @@
 
 package eu.esdihumboldt.hale.ui.service.population.internal;
 
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
@@ -33,7 +32,6 @@ import eu.esdihumboldt.hale.common.align.model.ChildContext;
 import eu.esdihumboldt.hale.common.align.model.EntityDefinition;
 import eu.esdihumboldt.hale.common.align.model.impl.TypeEntityDefinition;
 import eu.esdihumboldt.hale.common.instance.model.DataSet;
-import eu.esdihumboldt.hale.common.instance.model.Group;
 import eu.esdihumboldt.hale.common.instance.model.Instance;
 import eu.esdihumboldt.hale.common.instance.model.InstanceCollection;
 import eu.esdihumboldt.hale.common.instance.model.ResourceIterator;
@@ -44,6 +42,8 @@ import eu.esdihumboldt.hale.ui.common.service.population.PopulationService;
 import eu.esdihumboldt.hale.ui.common.service.population.impl.AbstractPopulationService;
 import eu.esdihumboldt.hale.ui.service.entity.EntityDefinitionService;
 import eu.esdihumboldt.hale.ui.service.entity.EntityDefinitionServiceListener;
+import eu.esdihumboldt.hale.ui.service.helper.population.IPopulationUpdater;
+import eu.esdihumboldt.hale.ui.service.helper.population.EntityPopulationCount;
 import eu.esdihumboldt.hale.ui.service.instance.InstanceService;
 import eu.esdihumboldt.hale.ui.service.instance.InstanceServiceAdapter;
 
@@ -52,7 +52,7 @@ import eu.esdihumboldt.hale.ui.service.instance.InstanceServiceAdapter;
  * 
  * @author Simon Templer
  */
-public class PopulationServiceImpl extends AbstractPopulationService {
+public class PopulationServiceImpl extends AbstractPopulationService implements IPopulationUpdater {
 
 	private static final Population NO_POPULATION = new Population() {
 
@@ -84,6 +84,8 @@ public class PopulationServiceImpl extends AbstractPopulationService {
 	private final Map<EntityDefinition, PopulationImpl> targetPopulation = new HashMap<EntityDefinition, PopulationImpl>();
 
 	private final EntityDefinitionService entityDefinitionService;
+
+	private static EntityPopulationCount populationCount;
 
 	/**
 	 * Create a population service instance.
@@ -173,6 +175,7 @@ public class PopulationServiceImpl extends AbstractPopulationService {
 
 		});
 
+		populationCount = new EntityPopulationCount(this);
 	}
 
 	/**
@@ -251,7 +254,7 @@ public class PopulationServiceImpl extends AbstractPopulationService {
 		for (TypeEntityDefinition def : typeDefinitions) {
 			if (def.getFilter() == null || def.getFilter().match(instance)) {
 				increase(def, 1);
-				addToPopulation(instance, def);
+				populationCount.addToPopulation(instance, def);
 			}
 		}
 	}
@@ -281,90 +284,21 @@ public class PopulationServiceImpl extends AbstractPopulationService {
 	}
 
 	/**
-	 * Count the population for the properties of the given group.
-	 * 
-	 * @param group the group
-	 * @param groupDef the group entity definition
+	 * @see eu.esdihumboldt.hale.ui.service.helper.population.IPopulationUpdater#increaseForEntity(eu.esdihumboldt.hale.common.align.model.EntityDefinition,
+	 *      int)
 	 */
-	private void addToPopulation(Group group, EntityDefinition groupDef) {
-		Iterable<? extends EntityDefinition> children = entityDefinitionService
-				.getChildren(groupDef);
-		if (children != null && children.iterator().hasNext()) {
-			for (EntityDefinition def : children) {
-				evaluateContext(group, def);
-			}
-		}
-		else {
-			evaluateContext(group, groupDef);
-		}
+	@Override
+	public void increaseForEntity(EntityDefinition def, int count) {
+		increase(def, count);
+
 	}
 
-	private void evaluateContext(Group group, EntityDefinition groupDef) {
-
-		List<ChildContext> path = groupDef.getPropertyPath();
-
-		if (path != null && !path.isEmpty()) {
-			ChildContext context = path.get(path.size() - 1);
-			Object[] values = group.getProperty(context.getChild().getName());
-			if (values != null) {
-				// apply the possible source contexts
-				if (context.getIndex() != null) {
-					// select only the item at the index
-					int index = context.getIndex();
-					if (index < values.length) {
-						values = new Object[] { values[index] };
-					}
-					else {
-						values = new Object[] {};
-					}
-				}
-				if (context.getCondition() != null) {
-					// select only values that match the condition
-					List<Object> matchedValues = new ArrayList<Object>();
-					for (Object value : values) {
-						if (AlignmentUtil.matchCondition(context.getCondition(), value, group)) {
-							matchedValues.add(value);
-						}
-					}
-					values = matchedValues.toArray();
-				}
-
-				if (context.getChild().getName().equals(groupDef.getDefinition().getName())) {
-					increase(groupDef, values.length);
-				}
-
-				for (Object value : values) {
-					if (value instanceof Group) {
-						addToPopulation((Group) value, groupDef);
-					}
-				}
-			}
-			else {
-				increase(groupDef, 0);
-			}
-		}
-	}
-
-	private void evaluateChildEntityDefinition(Group group, EntityDefinition groupDef,
-			List<ChildContext> path) {
-		if (path.size() == 1) {
-			evaluateContext(group, groupDef);
-		}
-		else {
-			ChildContext context = path.get(0);
-			List<ChildContext> subPath = path.subList(1, path.size());
-			Object[] values = group.getProperty(context.getChild().getName());
-			if (values != null) {
-				for (Object value : values) {
-					if (value instanceof Group) {
-						evaluateChildEntityDefinition((Group) value, groupDef, subPath);
-					}
-				}
-			}
-			else {
-				evaluateChildEntityDefinition(group, groupDef, subPath);
-			}
-		}
+	/**
+	 * @see eu.esdihumboldt.hale.ui.service.helper.population.IPopulationUpdater#getChildren(eu.esdihumboldt.hale.common.align.model.EntityDefinition)
+	 */
+	@Override
+	public Collection<? extends EntityDefinition> getChildren(EntityDefinition entityDef) {
+		return entityDefinitionService.getChildren(entityDef);
 	}
 
 	private void addNoneToPopulation(EntityDefinition groupDef, List<ChildContext> path) {
@@ -419,25 +353,6 @@ public class PopulationServiceImpl extends AbstractPopulationService {
 				pop.increaseOverall(values);
 			}
 
-		}
-	}
-
-	/**
-	 * Count population of {@link EntityDefinition}
-	 * 
-	 * @param entityDef a entity definition
-	 * @param instance an instance
-	 */
-	public void countPopulation(EntityDefinition entityDef, Instance instance) {
-		List<ChildContext> path = entityDef.getPropertyPath();
-		if (path == null || path.isEmpty()) {
-			if (entityDef.getFilter() == null || entityDef.getFilter().match(instance)) {
-				PopulationServiceImpl.this.increase(entityDef, 1);
-				PopulationServiceImpl.this.addToPopulation(instance, entityDef);
-			}
-		}
-		else {
-			PopulationServiceImpl.this.evaluateChildEntityDefinition(instance, entityDef, path);
 		}
 	}
 
@@ -515,8 +430,8 @@ public class PopulationServiceImpl extends AbstractPopulationService {
 					ResourceIterator<Instance> it = instances.iterator();
 					try {
 						while (it.hasNext()) {
-							PopulationServiceImpl.this.countPopulation(ccEntityDefinition,
-									it.next());
+							PopulationServiceImpl.populationCount
+									.countPopulation(ccEntityDefinition, it.next());
 						}
 					} finally {
 						it.close();
@@ -529,4 +444,5 @@ public class PopulationServiceImpl extends AbstractPopulationService {
 		}
 
 	}
+
 }
